+++ src/Rocksmith2014.DD/LevelCounter.fs
@@ -33,14 +33,28 @@ let predictLevelCount (path: int) (p: DataExtractor.PhraseData) =
 
     // Load model and predict the output
     let result =
-        lock lockObj (fun _ -> ConsumeModel.Predict(input))
+        lock lockObj (fun _ -> ConsumeModel.Predict input)
 
     let levels =
-        Math.Round(float result.Score)
+        round result.Score
         |> int
     Math.Clamp(levels, 2, 30)
 
+let private getRepeatedNotePercent (phraseData: DataExtractor.PhraseData) =
+    float (phraseData.RepeatedNotes + phraseData.RepeatedChords)
+    /
+    float (phraseData.NoteCount + phraseData.ChordCount)
+
 let getSimpleLevelCount (phraseData: DataExtractor.PhraseData)
                         (divisionMap: DivisionMap) =
+    let baseCount = divisionMap.Count
+
+    // Try to prevent inflated level count for phrases that are mostly repeated notes
+    let levelCount =
+        if baseCount > 15 && getRepeatedNotePercent phraseData > 0.8 then
+            baseCount / 2
+        else
+            baseCount
+
     let minLevels = max 2 phraseData.MaxChordStrings
-    Math.Clamp(divisionMap.Count, minLevels, 30)
+    Math.Clamp(levelCount, minLevels, 30)