+++ tests/Rocksmith2014.DD.Tests/GeneratorTests.fs
@@ -4,7 +4,9 @@ open Expecto
 open Rocksmith2014.XML
 open Rocksmith2014.DD
 
-let config = { PhraseSearch = WithThreshold 85 }
+let config =
+    { PhraseSearch = WithThreshold 85
+      LevelCountGeneration = LevelCountGeneration.Simple }
 
 [<Tests>]
 let ddGeneratorTests =